requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
django-filter~=24.3
Pillow==10.1.0
django-redis==5.4.0
redis==5.0.1
python-json-logger==2.0.7
